--- .idea/modules/informationserver.iml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module external.linked.project.id="informationserver" external.linked.project.path="$MODULE_DIR$/../../informationserver" external.root.project.path="$MODULE_DIR$/../../informationserver" external.system.id="GRADLE" external.system.module.group="com.clouway" external.system.module.version="1.0-SNAPSHOT" type="JAVA_MODULE" version="4">
-  <component name="NewModuleRootManager" inherit-compiler-output="true">
-    <exclude-output />
-    <content url="file://$MODULE_DIR$/../../informationserver">
-      <excludeFolder url="file://$MODULE_DIR$/../../informationserver/.gradle" />
-      <excludeFolder url="file://$MODULE_DIR$/../../informationserver/build" />
-    </content>
-    <orderEntry type="inheritedJdk" />
-    <orderEntry type="sourceFolder" forTests="false" />
-  </component>
-</module>
