--- .idea/modules/informationserver_main.iml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module external.linked.project.id="informationserver:main" external.linked.project.path="$MODULE_DIR$/../../informationserver" external.root.project.path="$MODULE_DIR$/../../informationserver" external.system.id="GRADLE" external.system.module.group="com.clouway" external.system.module.type="sourceSet" external.system.module.version="1.0-SNAPSHOT" type="JAVA_MODULE" version="4">
-  <component name="NewModuleRootManager" LANGUAGE_LEVEL="JDK_1_8" inherit-compiler-output="false">
-    <output url="file://$MODULE_DIR$/../../informationserver/build/classes/main" />
-    <exclude-output />
-    <content url="file://$MODULE_DIR$/../../informationserver/src/main">
-      <sourceFolder url="file://$MODULE_DIR$/../../informationserver/src/main/java" isTestSource="false" />
-      <sourceFolder url="file://$MODULE_DIR$/../../informationserver/src/main/resources" type="java-resource" />
-    </content>
-    <orderEntry type="inheritedJdk" />
-    <orderEntry type="sourceFolder" forTests="false" />
-  </component>
-</module>
