--- .idea/modules/informationserver_test.iml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module external.linked.project.id="informationserver:test" external.linked.project.path="$MODULE_DIR$/../../informationserver" external.root.project.path="$MODULE_DIR$/../../informationserver" external.system.id="GRADLE" external.system.module.group="com.clouway" external.system.module.type="sourceSet" external.system.module.version="1.0-SNAPSHOT" type="JAVA_MODULE" version="4">
-  <component name="NewModuleRootManager" LANGUAGE_LEVEL="JDK_1_8" inherit-compiler-output="false">
-    <output-test url="file://$MODULE_DIR$/../../informationserver/build/classes/test" />
-    <exclude-output />
-    <content url="file://$MODULE_DIR$/../../informationserver/src/test">
-      <sourceFolder url="file://$MODULE_DIR$/../../informationserver/src/test/java" isTestSource="true" />
-      <sourceFolder url="file://$MODULE_DIR$/../../informationserver/src/test/resources" type="java-test-resource" />
-    </content>
-    <orderEntry type="inheritedJdk" />
-    <orderEntry type="sourceFolder" forTests="false" />
-    <orderEntry type="module" module-name="informationserver_main" />
-    <orderEntry type="library" name="Gradle: org.jmock:jmock-junit4:2.8.2" level="project" />
-    <orderEntry type="library" name="Gradle: org.hamcrest:hamcrest-core:1.3" level="project" />
-    <orderEntry type="library" name="Gradle: org.jmock:jmock:2.8.2" level="project" />
-    <orderEntry type="library" name="Gradle: org.jmock:jmock-testjar:2.8.2" level="project" />
-    <orderEntry type="library" name="Gradle: cglib:cglib:3.2.0" level="project" />
-    <orderEntry type="library" name="Gradle: org.ow2.asm:asm:5.0.4" level="project" />
-    <orderEntry type="library" name="Gradle: org.beanshell:bsh:1.3.0" level="project" />
-    <orderEntry type="library" name="Gradle: org.hamcrest:hamcrest-library:1.3" level="project" />
-    <orderEntry type="library" name="Gradle: org.apache.ant:ant:1.9.4" level="project" />
-    <orderEntry type="library" name="Gradle: org.apache.ant:ant-launcher:1.9.4" level="project" />
-    <orderEntry type="library" name="Gradle: junit:junit:4.12" level="project" />
-  </component>
-  <component name="TestModuleProperties" production-module="informationserver_main" />
-</module>
