+++ clientcomunicationserver/.gradle/2.13/taskArtifacts/cache.properties
