+++ clientcomunicationserver/.gradle/2.13/taskArtifacts/fileHashes.bin
