+++ clientcomunicationserver/.gradle/2.13/taskArtifacts/fileSnapshots.bin
