+++ clientcomunicationserver/.gradle/2.13/taskArtifacts/taskArtifacts.bin
