+++ clientcomunicationserver/src/main/java/com/clouway/clientcomunicationserver/ClientInputMonitor.java
@@ -1,4 +1,4 @@
-package com.clouway.informationserver;
+package com.clouway.clientcomunicationserver;
 
 import java.io.BufferedReader;
 import java.io.IOException;
