+++ clientcomunicationserver/src/main/java/com/clouway/clientcomunicationserver/Main.java
@@ -1,4 +1,4 @@
-package com.clouway.informationserver;
+package com.clouway.clientcomunicationserver;
 
 import java.io.IOException;
 import java.util.LinkedList;
