+++ clientcomunicationserver/src/test/java/com/clouway/clientcomunicationserver/ServerReceptionistTest.java
@@ -1,4 +1,4 @@
-package com.clouway.informationserver;
+package com.clouway.clientcomunicationserver;
 
 import org.junit.Test;
